fastapi/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
snowflake-sqlalchemy
